# src/main/resources/openapi/account/account-api.yaml
openapi: 3.1.0
info:
  title: WealthPay Account API
  version: 1.0.0
  description: API for account management in WealthPay.

servers:
  - url: http://localhost:8080
    description: Local dev

paths:
  /accounts:
    post:
      tags:
        - Account
      summary: Open new bank account
      operationId: openAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenAccountRequest'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAccountResponse'

  /accounts/{id}:
    get:
      tags:
        - Account
      summary: Get the details of an account
      operationId: getAccountById
      parameters:
        - name: id
          in: path
          required: true
          description: Unique account identifier (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
  /accounts/{id}/deposits:
    post:
      tags:
        - Account Transaction
      summary: Deposit money into an account
      operationId: creditAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditAccountRequest'
      parameters:
        - name: id
          in: path
          required: true
          description: Unique account identifier (UUID)
          schema:
            type: string
            format: uuid
        - name: Transaction-Id
          in: header
          required: true
          description: Unique transaction identifier (UUID) for Idempotency
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transaction status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'

  /accounts/{id}/withdrawals:
    post:
      tags:
        - Account Transaction
      summary: Withdraw money from an account
      operationId: debitAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebitAccountRequest'
      parameters:
        - name: id
          in: path
          required: true
          description: Unique account identifier (UUID)
          schema:
            type: string
            format: uuid
        - name: Transaction-Id
          in: header
          required: true
          description: Unique transaction identifier (UUID) for Idempotency
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transaction status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'

  /accounts/{id}/reservations:
    post:
      tags:
        - Account Reservation
      summary: Reserve money for future operations
      operationId: reserveFunds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReserveFundsRequest'
      parameters:
        - name: id
          in: path
          required: true
          description: Unique account identifier (UUID)
          schema:
            type: string
            format: uuid
        - name: Transaction-Id
          in: header
          required: true
          description: Unique transaction identifier (UUID) for Idempotency
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reservation Status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReserveFundsResponse'

  /accounts/{id}/reservations/{reservationId}/capture:
    post:
      tags:
        - Account Reservation
      summary: Capture reservation
      operationId: captureReservation
      parameters:
        - name: id
          in: path
          required: true
          description: Unique account identifier (UUID)
          schema:
            type: string
            format: uuid
        - name: reservationId
          in: path
          required: true
          description: Unique reservation identifier (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reservation Status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureReservationResponse'

  /accounts/{id}/reservations/{reservationId}/cancel:
    post:
      tags:
        - Account Reservation
      summary: Cancel reservation
      operationId: cancelReservation
      parameters:
        - name: id
          in: path
          required: true
          description: Unique account identifier (UUID)
          schema:
            type: string
            format: uuid
        - name: reservationId
          in: path
          required: true
          description: Unique reservation identifier (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reservation Status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelReservationResponse'
components:
  schemas:
    OpenAccountRequest:
      type: object
      required:
        - accountCurrency
        - initialAmount
        - initialAmountCurrency
      properties:
        accountCurrency:
          description: Currency of the account (base currency of all operations)
          $ref: '#/components/schemas/SupportedCurrency'
        initialAmount:
          type: number
          description: Initial money credited when opening the account
          minimum: 0
        initialAmountCurrency:
          description: Currency of the initial deposit
          $ref: '#/components/schemas/SupportedCurrency'

    CreditAccountRequest:
      type: object
      required:
        - currency
        - amount
      properties:
        currency:
          description: Deposit currency
          $ref: '#/components/schemas/SupportedCurrency'
        amount:
          type: number
          description: Deposit amount
          exclusiveMinimum: 0

    DebitAccountRequest:
      type: object
      required:
        - currency
        - amount
      properties:
        currency:
          description: Withdrawal currency
          $ref: '#/components/schemas/SupportedCurrency'
        amount:
          type: number
          description: Withdrawal amount
          exclusiveMinimum: 0

    OpenAccountResponse:
      type: object
      required:
        - accountId
      properties:
        accountId:
          type: string
          format: uuid
          description: Unique identifier of the created account

    AccountResponse:
      type: object
      required:
        - id
        - currency
        - balanceAmount
        - reservedAmount
        - status
      properties:
        id:
          type: string
          format: uuid
        balanceAmount:
          type: number
          description: Current balance of the account
        currency:
          $ref: '#/components/schemas/SupportedCurrency'
          description: Currency of the account
        reservedAmount:
          type: number
          description: Amount reserved for future operations
        status:
          $ref: '#/components/schemas/AccountStatus'
          description: Current status of the account


    SupportedCurrency:
      type: string
      description: ISO 4217 currency code (e.g. USD, EUR, CHF)
      enum:
        - USD
        - AED
        - AUD
        - CAD
        - CHF
        - CNH
        - CZK
        - DKK
        - EUR
        - GBP
        - HKD
        - HUF
        - ILS
        - JPY
        - KRW
        - MXN
        - MYR
        - NOK
        - NZD
        - PLN
        - SAR
        - SEK
        - SGD
        - TRY
        - TWD
        - ZAR

    AccountStatus:
      type: string
      description: Current status of the account
      enum:
        - OPENED
        - CLOSED

    TransactionResponse:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/TransactionStatus'

    TransactionStatus:
      type: string
      description: Transaction Status
      enum:
        - NO_EFFECT
        - COMMITTED

    ReserveFundsRequest:
      type: object
      required:
        - currency
        - amount
      properties:
        currency:
          description: Reservation currency
          $ref: '#/components/schemas/SupportedCurrency'
        amount:
          type: number
          description: Reservation amount
          exclusiveMinimum: 0

    ReserveFundsResponse:
      type: object
      required:
        - status
        - reservationId
      properties:
        status:
          $ref: '#/components/schemas/ReserveFundsStatus'
        reservationId:
          type: string
          format: uuid

    ReserveFundsStatus:
      type: string
      description: Reservation Status
      enum:
        - RESERVED
        - NO_EFFECT

    CaptureReservationResponse:
      type: object
      required:
        - accountId
        - reservationId
        - status
      properties:
        accountId:
          type: string
          format: uuid
        reservationId:
          type: string
          format: uuid
        amount:
          type: number
        currency:
          $ref: '#/components/schemas/SupportedCurrency'
        status:
          $ref: '#/components/schemas/ReservationCaptureStatus'


    ReservationCaptureStatus:
      type: string
      description: Reservation Capture Status
      enum:
        - CAPTURED
        - NO_EFFECT

    CancelReservationResponse:
      type: object
      required:
        - accountId
        - reservationId
        - status
      properties:
        accountId:
          type: string
          format: uuid
        reservationId:
          type: string
          format: uuid
        amount:
          type: number
        currency:
          $ref: '#/components/schemas/SupportedCurrency'
        status:
          $ref: '#/components/schemas/ReservationCanceledStatus'

    ReservationCanceledStatus:
      type: string
      description: Reservation Cancel Status
      enum:
        - CANCELED
        - NO_EFFECT